remove legacy codes
