+++ CMakeLists.txt
@@ -45,15 +45,15 @@ project(ailoy C CXX)
 
 set(AILOY_WITH_TEST ON CACHE BOOL "Build test")
 if(DEFINED EMSCRIPTEN)
-    set(AILOY_WITH_MLC_LLM OFF CACHE BOOL "Enable mlc-llm(tvm) support")
+    set(AILOY_WITH_TVM OFF CACHE BOOL "Enable tvm support")
     set(AILOY_WITH_FAISS OFF CACHE BOOL "Enable faiss support")
     set(AILOY_USE_DUMMY_OPENMP ON CACHE BOOL "Use dummy OpenMP")
 else()
-    set(AILOY_WITH_MLC_LLM ON CACHE BOOL "Enable mlc-llm(tvm) support")
+    set(AILOY_WITH_TVM ON CACHE BOOL "Enable tvm support")
     set(AILOY_WITH_FAISS ON CACHE BOOL "Enable faiss support")
     set(AILOY_USE_DUMMY_OPENMP OFF CACHE BOOL "Use dummy openmp")
 endif()
-set(MLC_LLM_ROOT "" CACHE PATH "Root directory of mlc-llm. If AILOY_WITH_MLC_LLM is ON and this is empty, it will be downloaded automatically.")
+set(TVM_ROOT "" CACHE PATH "Root directory of mlc-llm. If AILOY_WITH_TVM is ON and this is empty, it will be downloaded automatically.")
 
 if(AILOY_USE_DUMMY_OPENMP)
     list(PREPEND CMAKE_MODULE_PATH "${PROJECT_SOURCE_DIR}/_deps/dummy_openmp/cmake")
