+++ vm/src/tvm/language_model.hpp
@@ -208,6 +208,6 @@ private:
 };
 
 component_or_error_t
-create_tvm_language_model_v2_component(std::shared_ptr<const value_t> inputs);
+create_tvm_language_model_component(std::shared_ptr<const value_t> inputs);
 
 } // namespace ailoy
